- change ValidateRange to return int instead of long
- remove nullable designator from RandomAccessStream.p_inputStream (optimization)
- convert long index values to int for benchmarks

+++ MetadataExtractor/IO/IRandomAccessStream.cs
@@ -29,6 +29,6 @@ namespace MetadataExtractor.IO
         void Seek(long index);
         byte[] ToArray(long index, int count);
 
-        long ValidateRange(long index, long bytesRequested);
+        int ValidateRange(long index, int bytesRequested);
     }
 }
